Created Lights.cpp, moved function definitions
